not a fix, but quick and dirty workaround for problem with Umlauten in OC6RC4
this one fixes core issue #6273.
Of course it's not meant as a real fix, but as something that has to be implemented the right way. I'm not so much in the code, that I can do it myself within a reasonable time, so I guess some of the core developer can do it much faster.