replace mrclay/minify with JShrink (#13052)
mrclay/minify is present only to provide the 'JSMin' minifier,
but this is non-free. JShrink is free, it's also a currently
maintained project following good development practices, and
not tied to a bundle of other junk like mrclay/minify. This
commit goes along with a commit to owncloud itself to switch
from using Assetic's JSMin filter to its JShrink filter.